+++ .riot/requirements/17e8568.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --no-annotate .riot/requirements/12739ce.in
+#    pip-compile --no-annotate .riot/requirements/17e8568.in
 #
 ai21==2.4.0
 ai21-tokenizer==0.9.1
@@ -25,6 +25,7 @@ fastavro==1.9.4
 filelock==3.14.0
 frozenlist==1.4.1
 fsspec==2024.5.0
+greenlet==3.0.3
 h11==0.14.0
 httpcore==1.0.5
 httpx==0.27.0
@@ -37,13 +38,13 @@ iniconfig==2.0.0
 jmespath==1.0.1
 jsonpatch==1.33
 jsonpointer==2.4
-langchain==0.1.20
+langchain==0.2.1
 langchain-aws==0.1.6
-langchain-community==0.0.38
-langchain-core==0.1.52
-langchain-openai==0.1.5
+langchain-community==0.2.1
+langchain-core==0.2.1
+langchain-openai==0.1.7
 langchain-pinecone==0.1.1
-langchain-text-splitters==0.0.2
+langchain-text-splitters==0.2.0
 langsmith==0.1.63
 marshmallow==3.21.2
 mock==5.1.0
@@ -51,7 +52,7 @@ multidict==6.0.5
 mypy-extensions==1.0.0
 numexpr==2.10.0
 numpy==1.26.4
-openai==1.12.0
+openai==1.30.4
 opentracing==2.4.0
 orjson==3.10.3
 packaging==23.2
@@ -68,7 +69,7 @@ pytest-randomly==3.15.0
 python-dateutil==2.9.0.post0
 pyyaml==6.0.1
 regex==2024.5.15
-requests==2.32.2
+requests==2.32.3
 s3transfer==0.10.1
 sentencepiece==0.2.0
 six==1.16.0
