+++ .riot/requirements/1ec5924.txt
@@ -1,31 +1,30 @@
 #
-# This file is autogenerated by pip-compile with Python 3.10
+# This file is autogenerated by pip-compile with Python 3.11
 # by the following command:
 #
-#    pip-compile --no-annotate .riot/requirements/a186e90.in
+#    pip-compile --no-annotate .riot/requirements/1ec5924.in
 #
 ai21==1.3.4
-aiohttp==3.9.3
+aiohttp==3.9.5
 aiosignal==1.3.1
-anyio==4.3.0
-async-timeout==4.0.3
+anyio==4.4.0
 attrs==23.2.0
 backoff==2.2.1
 certifi==2024.2.2
 charset-normalizer==3.3.2
-cohere==4.53
-coverage[toml]==7.4.3
+cohere==4.57
+coverage[toml]==7.5.3
 dataclasses-json==0.5.14
 dnspython==2.6.1
-exceptiongroup==1.2.0
+exceptiongroup==1.2.1
 fastavro==1.9.4
-filelock==3.13.1
+filelock==3.14.0
 frozenlist==1.4.1
-fsspec==2024.2.0
+fsspec==2024.5.0
 greenlet==3.0.3
-huggingface-hub==0.21.4
+huggingface-hub==0.23.2
 hypothesis==6.45.0
-idna==3.6
+idna==3.7
 importlib-metadata==6.11.0
 iniconfig==2.0.0
 jsonpatch==1.33
@@ -36,41 +35,40 @@ langchain-core==0.1.23
 langchainplus-sdk==0.0.4
 langsmith==0.0.87
 loguru==0.7.2
-marshmallow==3.21.1
+marshmallow==3.21.2
 mock==5.1.0
 multidict==6.0.5
 mypy-extensions==1.0.0
-numexpr==2.9.0
+numexpr==2.10.0
 numpy==1.26.4
 openai==0.27.8
 openapi-schema-pydantic==1.2.4
 opentracing==2.4.0
 packaging==23.2
 pinecone-client==2.2.4
-pluggy==1.4.0
+pluggy==1.5.0
 psutil==5.9.8
-pydantic==1.10.14
-pytest==8.1.1
+pydantic==1.10.15
+pytest==8.2.1
 pytest-asyncio==0.21.1
-pytest-cov==4.1.0
-pytest-mock==3.12.0
+pytest-cov==5.0.0
+pytest-mock==3.14.0
 pytest-randomly==3.15.0
 python-dateutil==2.9.0.post0
 pyyaml==6.0.1
-regex==2023.12.25
-requests==2.31.0
+regex==2024.5.15
+requests==2.32.3
 six==1.16.0
 sniffio==1.3.1
 sortedcontainers==2.4.0
-sqlalchemy==2.0.28
-tenacity==8.2.3
-tiktoken==0.6.0
-tomli==2.0.1
-tqdm==4.66.2
-typing-extensions==4.10.0
+sqlalchemy==2.0.30
+tenacity==8.3.0
+tiktoken==0.7.0
+tqdm==4.66.4
+typing-extensions==4.12.0
 typing-inspect==0.9.0
 urllib3==2.2.1
 vcrpy==6.0.1
 wrapt==1.16.0
 yarl==1.9.4
-zipp==3.17.0
+zipp==3.19.0
