+++ .riot/requirements/8297334.txt
@@ -1,8 +1,8 @@
 #
-# This file is autogenerated by pip-compile with Python 3.11
+# This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --no-annotate .riot/requirements/6251000.in
+#    pip-compile --no-annotate .riot/requirements/8297334.in
 #
 ai21==2.4.0
 ai21-tokenizer==0.9.1
@@ -10,6 +10,7 @@ aiohttp==3.9.5
 aiosignal==1.3.1
 annotated-types==0.7.0
 anyio==4.4.0
+async-timeout==4.0.3
 attrs==23.2.0
 boto3==1.34.114
 botocore==1.34.114
@@ -24,6 +25,7 @@ fastavro==1.9.4
 filelock==3.14.0
 frozenlist==1.4.1
 fsspec==2024.5.0
+greenlet==3.0.3
 h11==0.14.0
 httpcore==1.0.5
 httpx==0.27.0
@@ -31,6 +33,7 @@ httpx-sse==0.4.0
 huggingface-hub==0.23.2
 hypothesis==6.45.0
 idna==3.7
+importlib-metadata==7.1.0
 iniconfig==2.0.0
 jmespath==1.0.1
 jsonpatch==1.33
@@ -39,17 +42,17 @@ langchain==0.1.20
 langchain-aws==0.1.6
 langchain-community==0.0.38
 langchain-core==0.1.52
-langchain-openai==0.1.5
-langchain-pinecone==0.1.1
+langchain-openai==0.1.6
+langchain-pinecone==0.1.0
 langchain-text-splitters==0.0.2
-langsmith==0.1.63
+langsmith==0.1.58
 marshmallow==3.21.2
 mock==5.1.0
 multidict==6.0.5
 mypy-extensions==1.0.0
 numexpr==2.10.0
 numpy==1.26.4
-openai==1.12.0
+openai==1.30.3
 opentracing==2.4.0
 orjson==3.10.3
 packaging==23.2
@@ -66,7 +69,7 @@ pytest-randomly==3.15.0
 python-dateutil==2.9.0.post0
 pyyaml==6.0.1
 regex==2024.5.15
-requests==2.32.2
+requests==2.32.3
 s3transfer==0.10.1
 sentencepiece==0.2.0
 six==1.16.0
@@ -76,11 +79,14 @@ sqlalchemy==2.0.30
 tenacity==8.3.0
 tiktoken==0.7.0
 tokenizers==0.19.1
+tomli==2.0.1
 tqdm==4.66.4
-types-requests==2.32.0.20240523
+types-requests==2.31.0.6
+types-urllib3==1.26.25.14
 typing-extensions==4.12.0
 typing-inspect==0.9.0
-urllib3==2.2.1
+urllib3==1.26.18
 vcrpy==6.0.1
 wrapt==1.16.0
 yarl==1.9.4
+zipp==3.19.0
