+++ .riot/requirements/ee6f953.txt
@@ -1,0 +1,87 @@
+#
+# This file is autogenerated by pip-compile with Python 3.11
+# by the following command:
+#
+#    pip-compile --no-annotate .riot/requirements/ee6f953.in
+#
+ai21==2.4.0
+ai21-tokenizer==0.9.1
+aiohttp==3.9.5
+aiosignal==1.3.1
+annotated-types==0.7.0
+anyio==4.4.0
+attrs==23.2.0
+boto3==1.34.114
+botocore==1.34.114
+certifi==2024.2.2
+charset-normalizer==3.3.2
+cohere==5.5.3
+coverage[toml]==7.5.3
+dataclasses-json==0.6.6
+distro==1.9.0
+exceptiongroup==1.2.1
+fastavro==1.9.4
+filelock==3.14.0
+frozenlist==1.4.1
+fsspec==2024.5.0
+greenlet==3.0.3
+h11==0.14.0
+httpcore==1.0.5
+httpx==0.27.0
+httpx-sse==0.4.0
+huggingface-hub==0.23.2
+hypothesis==6.45.0
+idna==3.7
+iniconfig==2.0.0
+jmespath==1.0.1
+jsonpatch==1.33
+jsonpointer==2.4
+langchain==0.2.1
+langchain-aws==0.1.6
+langchain-community==0.2.1
+langchain-core==0.2.1
+langchain-openai==0.1.7
+langchain-pinecone==0.1.1
+langchain-text-splitters==0.2.0
+langsmith==0.1.63
+marshmallow==3.21.2
+mock==5.1.0
+multidict==6.0.5
+mypy-extensions==1.0.0
+numexpr==2.10.0
+numpy==1.26.4
+openai==1.30.4
+opentracing==2.4.0
+orjson==3.10.3
+packaging==23.2
+pinecone-client==3.2.2
+pluggy==1.5.0
+psutil==5.9.8
+pydantic==2.7.2
+pydantic-core==2.18.3
+pytest==8.2.1
+pytest-asyncio==0.21.1
+pytest-cov==5.0.0
+pytest-mock==3.14.0
+pytest-randomly==3.15.0
+python-dateutil==2.9.0.post0
+pyyaml==6.0.1
+regex==2024.5.15
+requests==2.32.3
+s3transfer==0.10.1
+sentencepiece==0.2.0
+six==1.16.0
+sniffio==1.3.1
+sortedcontainers==2.4.0
+sqlalchemy==2.0.30
+tenacity==8.3.0
+tiktoken==0.7.0
+tokenizers==0.19.1
+tqdm==4.66.4
+types-requests==2.32.0.20240523
+typing-extensions==4.12.0
+typing-inspect==0.9.0
+urllib3==2.2.1
+vcrpy==6.0.1
+wrapt==1.16.0
+yarl==1.9.4
