+++ tests/contrib/langchain/test_langchain_community.py
@@ -1,4 +1,3 @@
-import json
 from operator import itemgetter
 import os
 import re
@@ -10,12 +9,7 @@ import mock
 import pytest
 
 from ddtrace.internal.utils.version import parse_version
-from ddtrace.llmobs import LLMObs
 from tests.contrib.langchain.utils import get_request_vcr
-from tests.llmobs._utils import _expected_llmobs_llm_span_event
-from tests.llmobs._utils import _expected_llmobs_non_llm_span_event
-from tests.subprocesstest import SubprocessTestCase
-from tests.subprocesstest import run_in_subprocess
 from tests.utils import flaky
 from tests.utils import override_global_config
 
@@ -34,7 +28,6 @@ def request_vcr():
     yield get_request_vcr(subdirectory_name="langchain_community")
 
 
-@flaky(1735812000)
 @pytest.mark.parametrize("ddtrace_config_langchain", [dict(logs_enabled=True, log_prompt_completion_sample_rate=1.0)])
 def test_global_tags(
     ddtrace_config_langchain, langchain, langchain_openai, request_vcr, mock_metrics, mock_logs, mock_tracer
@@ -86,7 +79,6 @@ def test_global_tags(
         )
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot(ignores=["metrics.langchain.tokens.total_cost", "resource"])
 def test_openai_llm_sync(langchain, langchain_openai, request_vcr):
     llm = langchain_openai.OpenAI()
@@ -94,7 +86,6 @@ def test_openai_llm_sync(langchain, langchain_openai, request_vcr):
         llm.invoke("Can you explain what Descartes meant by 'I think, therefore I am'?")
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_openai_llm_sync_multiple_prompts(langchain, langchain_openai, request_vcr):
     llm = langchain_openai.OpenAI()
@@ -130,7 +121,6 @@ async def test_openai_llm_async_stream(langchain, langchain_openai, request_vcr)
         await llm.agenerate(["Why is Spongebob so bad at driving?"])
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot(ignores=["meta.error.stack", "resource"])
 def test_openai_llm_error(langchain, langchain_openai, request_vcr):
     import openai  # Imported here because the os env OPENAI_API_KEY needs to be set via langchain fixture before import
@@ -226,7 +216,6 @@ def test_llm_logs(
     mock_metrics.count.assert_not_called()
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_openai_chat_model_sync_call_langchain_openai(langchain, langchain_openai, request_vcr):
     chat = langchain_openai.ChatOpenAI(temperature=0, max_tokens=256)
@@ -234,7 +223,6 @@ def test_openai_chat_model_sync_call_langchain_openai(langchain, langchain_opena
         chat.invoke(input=[langchain.schema.HumanMessage(content="When do you use 'whom' instead of 'who'?")])
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_openai_chat_model_sync_generate(langchain, langchain_openai, request_vcr):
     chat = langchain_openai.ChatOpenAI(temperature=0, max_tokens=256)
@@ -328,7 +316,6 @@ async def test_openai_chat_model_async_stream(langchain, langchain_openai, reque
         await chat.agenerate([[langchain.schema.HumanMessage(content="What is the secret Krabby Patty recipe?")]])
 
 
-@flaky(1735812000)
 def test_chat_model_metrics(langchain, langchain_openai, request_vcr, mock_metrics, mock_logs, snapshot_tracer):
     chat = langchain_openai.ChatOpenAI(temperature=0, max_tokens=256)
     with request_vcr.use_cassette("openai_chat_completion_sync_call.yaml"):
@@ -356,7 +343,6 @@ def test_chat_model_metrics(langchain, langchain_openai, request_vcr, mock_metri
     mock_logs.assert_not_called()
 
 
-@flaky(1735812000)
 @pytest.mark.parametrize(
     "ddtrace_config_langchain",
     [dict(metrics_enabled=False, logs_enabled=True, log_prompt_completion_sample_rate=1.0)],
@@ -487,7 +473,6 @@ def test_embedding_logs(langchain_openai, ddtrace_config_langchain, request_vcr,
     mock_metrics.count.assert_not_called()
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_openai_math_chain_sync(langchain, langchain_openai, request_vcr):
     """
@@ -499,7 +484,6 @@ def test_openai_math_chain_sync(langchain, langchain_openai, request_vcr):
         chain.invoke("what is two raised to the fifty-fourth power?")
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot(token="tests.contrib.langchain.test_langchain_community.test_chain_invoke")
 def test_chain_invoke_dict_input(langchain, langchain_openai, request_vcr):
     prompt_template = "what is {base} raised to the fifty-fourth power?"
@@ -509,7 +493,6 @@ def test_chain_invoke_dict_input(langchain, langchain_openai, request_vcr):
         chain.invoke(input={"base": "two"})
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot(token="tests.contrib.langchain.test_langchain_community.test_chain_invoke")
 def test_chain_invoke_str_input(langchain, langchain_openai, request_vcr):
     prompt_template = "what is {base} raised to the fifty-fourth power?"
@@ -597,7 +580,6 @@ def test_openai_sequential_chain(langchain, langchain_openai, request_vcr):
         sequential_chain.invoke({"text": input_text, "style": "a 90s rapper"})
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_openai_sequential_chain_with_multiple_llm_sync(langchain, langchain_openai, request_vcr):
     template = """Paraphrase this text:
@@ -677,7 +659,6 @@ async def test_openai_sequential_chain_with_multiple_llm_async(langchain, langch
         await sequential_chain.ainvoke({"input_text": input_text})
 
 
-@flaky(1735812000)
 @pytest.mark.parametrize(
     "ddtrace_config_langchain",
     [dict(metrics_enabled=False, logs_enabled=True, log_prompt_completion_sample_rate=1.0)],
@@ -928,7 +909,7 @@ def test_vectorstore_logs(
     mock_metrics.count.assert_not_called()
 
 
-@pytest.mark.snapshot(ignores=["metrics.langchain.tokens.total_cost", "resource"])
+@pytest.mark.snapshot(ignores=["metrics.langchain.tokens.total_cost", "meta.http.useragent", "resource"])
 def test_openai_integration(langchain, request_vcr, ddtrace_run_python_code_in_subprocess):
     env = os.environ.copy()
     pypath = [os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(__file__))))]
@@ -959,7 +940,7 @@ with get_request_vcr(subdirectory_name="langchain_community").use_cassette("open
     assert err == b""
 
 
-@pytest.mark.snapshot
+@pytest.mark.snapshot(ignores=["meta.http.useragent"])
 @pytest.mark.parametrize("schema_version", [None, "v0", "v1"])
 @pytest.mark.parametrize("service_name", [None, "mysvc"])
 def test_openai_service_name(
@@ -1001,7 +982,6 @@ with get_request_vcr(subdirectory_name="langchain_community").use_cassette("open
     assert err == b""
 
 
-@flaky(1735812000)
 @pytest.mark.parametrize(
     "ddtrace_config_langchain",
     [dict(metrics_enabled=False, logs_enabled=True, log_prompt_completion_sample_rate=1.0)],
@@ -1120,7 +1100,6 @@ def test_embedding_logs_when_response_not_completed(
     )
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_lcel_chain_simple(langchain_core, langchain_openai, request_vcr):
     prompt = langchain_core.prompts.ChatPromptTemplate.from_messages(
@@ -1133,7 +1112,6 @@ def test_lcel_chain_simple(langchain_core, langchain_openai, request_vcr):
         chain.invoke({"input": "how can langsmith help with testing?"})
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_lcel_chain_complicated(langchain_core, langchain_openai, request_vcr):
     prompt = langchain_core.prompts.ChatPromptTemplate.from_template(
@@ -1176,7 +1154,7 @@ async def test_lcel_chain_simple_async(langchain_core, langchain_openai, request
         await chain.ainvoke({"input": "how can langsmith help with testing?"})
 
 
-@flaky(1735812000)
+@flaky(1735812000, reason="batch() is non-deterministic in which order it processes inputs")
 @pytest.mark.snapshot
 @pytest.mark.skipif(sys.version_info >= (3, 11, 0), reason="Python <3.11 test")
 def test_lcel_chain_batch(langchain_core, langchain_openai, request_vcr):
@@ -1193,7 +1171,7 @@ def test_lcel_chain_batch(langchain_core, langchain_openai, request_vcr):
         chain.batch(inputs=["chickens", "pigs"])
 
 
-@flaky(1735812000)
+@flaky(1735812000, reason="batch() is non-deterministic in which order it processes inputs")
 @pytest.mark.snapshot
 @pytest.mark.skipif(sys.version_info < (3, 11, 0), reason="Python 3.11+ required")
 def test_lcel_chain_batch_311(langchain_core, langchain_openai, request_vcr):
@@ -1210,7 +1188,6 @@ def test_lcel_chain_batch_311(langchain_core, langchain_openai, request_vcr):
         chain.batch(inputs=["chickens", "pigs"])
 
 
-@flaky(1735812000)
 @pytest.mark.snapshot
 def test_lcel_chain_nested(langchain_core, langchain_openai, request_vcr):
     """
@@ -1234,7 +1211,7 @@ def test_lcel_chain_nested(langchain_core, langchain_openai, request_vcr):
         complete_chain.invoke({"person": "Spongebob Squarepants", "language": "Spanish"})
 
 
-@flaky(1735812000)
+@flaky(1735812000, reason="batch() is non-deterministic in which order it processes inputs")
 @pytest.mark.asyncio
 @pytest.mark.snapshot
 async def test_lcel_chain_batch_async(langchain_core, langchain_openai, request_vcr):
@@ -1261,541 +1238,3 @@ def test_lcecl_chain_non_dict_input(langchain_core):
     sequence = add_one | multiply_two
 
     sequence.invoke(1)
-
-
-@pytest.mark.parametrize(
-    "ddtrace_global_config",
-    [dict(_llmobs_enabled=True, _llmobs_sample_rate=1.0, _llmobs_ml_app="langchain_community_test")],
-)
-class TestLLMObsLangchain:
-    @staticmethod
-    def _expected_llmobs_chain_calls(trace, expected_spans_data: list):
-        expected_llmobs_writer_calls = [mock.call.start()]
-
-        for idx, span in enumerate(trace):
-            kind, kwargs = expected_spans_data[idx]
-            expected_span_event = None
-            if kind == "chain":
-                expected_span_event = TestLLMObsLangchain._expected_llmobs_chain_call(span, **kwargs)
-            else:
-                expected_span_event = TestLLMObsLangchain._expected_llmobs_llm_call(span, **kwargs)
-
-            expected_llmobs_writer_calls += [mock.call.enqueue(expected_span_event)]
-
-        return expected_llmobs_writer_calls
-
-    @staticmethod
-    def _expected_llmobs_chain_call(span, input_parameters=None, input_value=None, output_value=None):
-        return _expected_llmobs_non_llm_span_event(
-            span,
-            span_kind="workflow",
-            parameters=input_parameters,
-            input_value=input_value,
-            output_value=output_value,
-            tags={
-                "ml_app": "langchain_community_test",
-            },
-            integration="langchain",
-        )
-
-    @staticmethod
-    def _expected_llmobs_llm_call(span, provider="openai", input_roles=[None], output_role=None):
-        input_meta = [{"content": mock.ANY} for _ in input_roles]
-        for idx, role in enumerate(input_roles):
-            if role is not None:
-                input_meta[idx]["role"] = role
-
-        output_meta = {"content": mock.ANY}
-        if output_role is not None:
-            output_meta["role"] = output_role
-
-        temperature_key = "temperature"
-        if provider == "huggingface_hub":
-            max_tokens_key = "model_kwargs.max_tokens"
-            temperature_key = "model_kwargs.temperature"
-        elif provider == "ai21":
-            max_tokens_key = "maxTokens"
-        else:
-            max_tokens_key = "max_tokens"
-
-        metadata = {}
-        temperature = span.get_tag(f"langchain.request.{provider}.parameters.{temperature_key}")
-        max_tokens = span.get_tag(f"langchain.request.{provider}.parameters.{max_tokens_key}")
-        if temperature is not None:
-            metadata["temperature"] = float(temperature)
-        if max_tokens is not None:
-            metadata["max_tokens"] = int(max_tokens)
-
-        return _expected_llmobs_llm_span_event(
-            span,
-            model_name=span.get_tag("langchain.request.model"),
-            model_provider=span.get_tag("langchain.request.provider"),
-            input_messages=input_meta,
-            output_messages=[output_meta],
-            metadata=metadata,
-            token_metrics={},
-            tags={
-                "ml_app": "langchain_community_test",
-            },
-            integration="langchain",
-        )
-
-    @classmethod
-    def _test_llmobs_llm_invoke(
-        cls,
-        provider,
-        generate_trace,
-        request_vcr,
-        mock_llmobs_span_writer,
-        mock_tracer,
-        cassette_name,
-        input_roles=[None],
-        output_role=None,
-    ):
-        LLMObs.disable()
-        LLMObs.enable(_tracer=mock_tracer, integrations_enabled=False)  # only want langchain patched
-
-        with request_vcr.use_cassette(cassette_name):
-            generate_trace("Can you explain what an LLM chain is?")
-        span = mock_tracer.pop_traces()[0][0]
-
-        expected_llmons_writer_calls = [
-            mock.call.start(),
-            mock.call.enqueue(
-                cls._expected_llmobs_llm_call(
-                    span,
-                    provider=provider,
-                    input_roles=input_roles,
-                    output_role=output_role,
-                )
-            ),
-        ]
-
-        assert mock_llmobs_span_writer.enqueue.call_count == 1
-        mock_llmobs_span_writer.assert_has_calls(expected_llmons_writer_calls)
-
-    @classmethod
-    def _test_llmobs_chain_invoke(
-        cls,
-        generate_trace,
-        request_vcr,
-        mock_llmobs_span_writer,
-        mock_tracer,
-        cassette_name,
-        expected_spans_data=[("llm", {"provider": "openai", "input_roles": [None], "output_role": None})],
-    ):
-        # disable the service before re-enabling it, as it was enabled in another test
-        LLMObs.disable()
-        LLMObs.enable(_tracer=mock_tracer, integrations_enabled=False)  # only want langchain patched
-
-        with request_vcr.use_cassette(cassette_name):
-            generate_trace("Can you explain what an LLM chain is?")
-        trace = mock_tracer.pop_traces()[0]
-
-        expected_llmobs_writer_calls = cls._expected_llmobs_chain_calls(
-            trace=trace, expected_spans_data=expected_spans_data
-        )
-        assert mock_llmobs_span_writer.enqueue.call_count == len(expected_spans_data)
-        mock_llmobs_span_writer.assert_has_calls(expected_llmobs_writer_calls)
-
-    @flaky(1735812000)
-    def test_llmobs_openai_llm(self, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr):
-        llm = langchain_openai.OpenAI()
-
-        self._test_llmobs_llm_invoke(
-            generate_trace=llm.invoke,
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="openai_completion_sync.yaml",
-            provider="openai",
-        )
-
-    def test_llmobs_cohere_llm(self, langchain_community, mock_llmobs_span_writer, mock_tracer, request_vcr):
-        llm = langchain_community.llms.Cohere(model="cohere.command-light-text-v14")
-
-        self._test_llmobs_llm_invoke(
-            generate_trace=llm.invoke,
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="cohere_completion_sync.yaml",
-            provider="cohere",
-        )
-
-    def test_llmobs_ai21_llm(self, langchain_community, mock_llmobs_span_writer, mock_tracer, request_vcr):
-        llm = langchain_community.llms.AI21()
-
-        self._test_llmobs_llm_invoke(
-            generate_trace=llm.invoke,
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="ai21_completion_sync.yaml",
-            provider="ai21",
-        )
-
-    @flaky(1735812000)
-    def test_llmobs_openai_chat_model(self, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr):
-        chat = langchain_openai.ChatOpenAI(temperature=0, max_tokens=256)
-
-        self._test_llmobs_llm_invoke(
-            generate_trace=lambda prompt: chat.invoke([langchain.schema.HumanMessage(content=prompt)]),
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="openai_chat_completion_sync_call.yaml",
-            provider="openai",
-            input_roles=["user"],
-            output_role="assistant",
-        )
-
-    @flaky(1735812000)
-    def test_llmobs_openai_chat_model_custom_role(
-        self, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr
-    ):
-        chat = langchain_openai.ChatOpenAI(temperature=0, max_tokens=256)
-
-        self._test_llmobs_llm_invoke(
-            generate_trace=lambda prompt: chat.invoke([langchain.schema.ChatMessage(content=prompt, role="custom")]),
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="openai_chat_completion_sync_call.yaml",
-            provider="openai",
-            input_roles=["custom"],
-            output_role="assistant",
-        )
-
-    @flaky(1735812000)
-    def test_llmobs_chain(self, langchain_core, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr):
-        prompt = langchain_core.prompts.ChatPromptTemplate.from_messages(
-            [("system", "You are world class technical documentation writer."), ("user", "{input}")]
-        )
-        llm = langchain_openai.OpenAI()
-
-        chain = prompt | llm
-
-        expected_output = (
-            "\nSystem: Langsmith can help with testing in several ways. "
-            "First, it can generate automated tests based on your technical documentation, "
-            "ensuring that your code matches the documented specifications. "
-            "This can save you time and effort in testing your code manually. "
-            "Additionally, Langsmith can also analyze your technical documentation for completeness and accuracy, "
-            "helping you identify any potential gaps or errors before testing begins. "
-            "Finally, Langsmith can assist with creating test cases and scenarios based on your documentation, "
-            "making the testing process more efficient and effective."
-        )
-
-        self._test_llmobs_chain_invoke(
-            generate_trace=lambda prompt: chain.invoke({"input": prompt}),
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="lcel_openai_chain_call.yaml",
-            expected_spans_data=[
-                (
-                    "chain",
-                    {
-                        "input_value": json.dumps([{"input": "Can you explain what an LLM chain is?"}]),
-                        "output_value": expected_output,
-                    },
-                ),
-                ("llm", {"provider": "openai", "input_roles": [None], "output_role": None}),
-            ],
-        )
-
-    def test_llmobs_chain_nested(
-        self, langchain_core, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr
-    ):
-        prompt1 = langchain_core.prompts.ChatPromptTemplate.from_template("what is the city {person} is from?")
-        prompt2 = langchain_core.prompts.ChatPromptTemplate.from_template(
-            "what country is the city {city} in? respond in {language}"
-        )
-
-        model = langchain_openai.ChatOpenAI()
-
-        chain1 = prompt1 | model | langchain_core.output_parsers.StrOutputParser()
-        chain2 = prompt2 | model | langchain_core.output_parsers.StrOutputParser()
-
-        complete_chain = {"city": chain1, "language": itemgetter("language")} | chain2
-
-        self._test_llmobs_chain_invoke(
-            generate_trace=lambda inputs: complete_chain.invoke(
-                {"person": "Spongebob Squarepants", "language": "Spanish"}
-            ),
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="lcel_openai_chain_nested.yaml",
-            expected_spans_data=[
-                (
-                    "chain",
-                    {
-                        "input_value": json.dumps([{"person": "Spongebob Squarepants", "language": "Spanish"}]),
-                        "output_value": mock.ANY,
-                    },
-                ),
-                (
-                    "chain",
-                    {
-                        "input_value": json.dumps([{"person": "Spongebob Squarepants", "language": "Spanish"}]),
-                        "output_value": mock.ANY,
-                    },
-                ),
-                ("llm", {"provider": "openai", "input_roles": ["user"], "output_role": "assistant"}),
-                ("llm", {"provider": "openai", "input_roles": ["user"], "output_role": "assistant"}),
-            ],
-        )
-
-    @pytest.mark.skipif(sys.version_info >= (3, 11, 0), reason="Python <3.11 required")
-    def test_llmobs_chain_batch(
-        self, langchain_core, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr
-    ):
-        prompt = langchain_core.prompts.ChatPromptTemplate.from_template("Tell me a short joke about {topic}")
-        output_parser = langchain_core.output_parsers.StrOutputParser()
-        model = langchain_openai.ChatOpenAI()
-        chain = {"topic": langchain_core.runnables.RunnablePassthrough()} | prompt | model | output_parser
-
-        self._test_llmobs_chain_invoke(
-            generate_trace=lambda inputs: chain.batch(inputs=["chickens", "pigs"]),
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="lcel_openai_chain_batch.yaml",
-            expected_spans_data=[
-                (
-                    "chain",
-                    {
-                        "input_value": json.dumps(["chickens", "pigs"]),
-                        "output_value": mock.ANY,
-                    },
-                ),
-                ("llm", {"provider": "openai", "input_roles": ["user"], "output_role": "assistant"}),
-                ("llm", {"provider": "openai", "input_roles": ["user"], "output_role": "assistant"}),
-            ],
-        )
-
-    @flaky(1735812000)
-    def test_llmobs_chain_schema_io(
-        self, langchain_core, langchain_openai, mock_llmobs_span_writer, mock_tracer, request_vcr
-    ):
-        model = langchain_openai.ChatOpenAI()
-        prompt = langchain_core.prompts.ChatPromptTemplate.from_messages(
-            [
-                ("system", "You're an assistant who's good at {ability}. Respond in 20 words or fewer"),
-                langchain_core.prompts.MessagesPlaceholder(variable_name="history"),
-                ("human", "{input}"),
-            ]
-        )
-
-        chain = prompt | model
-
-        self._test_llmobs_chain_invoke(
-            generate_trace=lambda inputs: chain.invoke(
-                {
-                    "ability": "world capitals",
-                    "history": [
-                        langchain.schema.HumanMessage(content="Can you be my science teacher instead?"),
-                        langchain.schema.AIMessage(content="Yes"),
-                    ],
-                    "input": "What's the powerhouse of the cell?",
-                }
-            ),
-            request_vcr=request_vcr,
-            mock_llmobs_span_writer=mock_llmobs_span_writer,
-            mock_tracer=mock_tracer,
-            cassette_name="lcel_openai_chain_schema_io.yaml",
-            expected_spans_data=[
-                (
-                    "chain",
-                    {
-                        "input_value": json.dumps(
-                            [
-                                {
-                                    "ability": "world capitals",
-                                    "history": [
-                                        ["user", "Can you be my science teacher instead?"],
-                                        ["assistant", "Yes"],
-                                    ],
-                                    "input": "What's the powerhouse of the cell?",
-                                }
-                            ]
-                        ),
-                        "output_value": json.dumps(["assistant", "Mitochondria."]),
-                    },
-                ),
-                (
-                    "llm",
-                    {
-                        "provider": "openai",
-                        "input_roles": ["system", "user", "assistant", "user"],
-                        "output_role": "assistant",
-                    },
-                ),
-            ],
-        )
-
-
-class TestLangchainTraceStructureWithLlmIntegrations(SubprocessTestCase):
-    bedrock_env_config = dict(
-        AWS_ACCESS_KEY_ID="testing",
-        AWS_SECRET_ACCESS_KEY="testing",
-        AWS_SECURITY_TOKEN="testing",
-        AWS_SESSION_TOKEN="testing",
-        AWS_DEFAULT_REGION="us-east-1",
-        DD_LANGCHAIN_METRICS_ENABLED="false",
-        DD_API_KEY="<not-a-real-key>",
-    )
-
-    openai_env_config = dict(
-        OPENAI_API_KEY="testing",
-        DD_API_KEY="<not-a-real-key>",
-    )
-
-    def setUp(self):
-        patcher = mock.patch("ddtrace.llmobs._llmobs.LLMObsSpanWriter")
-        LLMObsSpanWriterMock = patcher.start()
-        mock_llmobs_span_writer = mock.MagicMock()
-        LLMObsSpanWriterMock.return_value = mock_llmobs_span_writer
-
-        self.mock_llmobs_span_writer = mock_llmobs_span_writer
-
-        super(TestLangchainTraceStructureWithLlmIntegrations, self).setUp()
-
-    def _assert_trace_structure_from_writer_call_args(self, span_kinds):
-        assert self.mock_llmobs_span_writer.enqueue.call_count == len(span_kinds)
-
-        calls = self.mock_llmobs_span_writer.enqueue.call_args_list
-
-        for span_kind, call in zip(span_kinds, calls):
-            call_args = call.args[0]
-
-            assert call_args["meta"]["span.kind"] == span_kind
-            if span_kind == "workflow":
-                assert len(call_args["meta"]["input"]["value"]) > 0
-                assert len(call_args["meta"]["output"]["value"]) > 0
-            elif span_kind == "llm":
-                assert len(call_args["meta"]["input"]["messages"]) > 0
-                assert len(call_args["meta"]["output"]["messages"]) > 0
-
-    def _call_bedrock_chat_model(self, ChatBedrock, HumanMessage):
-        chat = ChatBedrock(
-            model_id="amazon.titan-tg1-large",
-            model_kwargs={"max_tokens": 50, "temperature": 0},
-        )
-        messages = [HumanMessage(content="summarize the plot to the lord of the rings in a dozen words")]
-        with get_request_vcr(subdirectory_name="langchain_community").use_cassette("bedrock_amazon_chat_invoke.yaml"):
-            chat.invoke(messages)
-
-    def _call_bedrock_llm(self, Bedrock, ConversationChain, ConversationBufferMemory):
-        llm = Bedrock(
-            model_id="amazon.titan-tg1-large",
-            region_name="us-east-1",
-            model_kwargs={"temperature": 0, "topP": 0.9, "stopSequences": [], "maxTokens": 50},
-        )
-
-        conversation = ConversationChain(llm=llm, verbose=True, memory=ConversationBufferMemory())
-
-        with get_request_vcr(subdirectory_name="langchain_community").use_cassette("bedrock_amazon_invoke.yaml"):
-            conversation.predict(input="can you explain what Datadog is to someone not in the tech industry?")
-
-    def _call_openai_llm(self, OpenAI):
-        llm = OpenAI()
-        with get_request_vcr(subdirectory_name="langchain_community").use_cassette("openai_completion_sync.yaml"):
-            llm.invoke("Can you explain what Descartes meant by 'I think, therefore I am'?")
-
-    @run_in_subprocess(env_overrides=bedrock_env_config)
-    def test_llmobs_with_chat_model_bedrock_enabled(self):
-        from langchain_aws import ChatBedrock
-        from langchain_core.messages import HumanMessage
-
-        from ddtrace import patch
-        from ddtrace.llmobs import LLMObs
-
-        patch(langchain=True, botocore=True)
-        LLMObs.enable(ml_app="<ml-app-name>", integrations_enabled=False, agentless_enabled=True)
-
-        self._call_bedrock_chat_model(ChatBedrock, HumanMessage)
-
-        self._assert_trace_structure_from_writer_call_args(["workflow", "llm"])
-
-        LLMObs.disable()
-
-    @run_in_subprocess(env_overrides=bedrock_env_config)
-    def test_llmobs_with_chat_model_bedrock_disabled(self):
-        from langchain_aws import ChatBedrock
-        from langchain_core.messages import HumanMessage
-
-        from ddtrace import patch
-        from ddtrace.llmobs import LLMObs
-
-        patch(langchain=True)
-        LLMObs.enable(ml_app="<ml-app-name>", integrations_enabled=False, agentless_enabled=True)
-
-        self._call_bedrock_chat_model(ChatBedrock, HumanMessage)
-
-        self._assert_trace_structure_from_writer_call_args(["llm"])
-
-        LLMObs.disable()
-
-    @run_in_subprocess(env_overrides=bedrock_env_config)
-    def test_llmobs_with_llm_model_bedrock_enabled(self):
-        from langchain.chains import ConversationChain
-        from langchain.memory import ConversationBufferMemory
-        from langchain_community.llms import Bedrock
-
-        from ddtrace import patch
-        from ddtrace.llmobs import LLMObs
-
-        patch(langchain=True, botocore=True)
-        LLMObs.enable(ml_app="<ml-app-name>", integrations_enabled=False, agentless_enabled=True)
-        self._call_bedrock_llm(Bedrock, ConversationChain, ConversationBufferMemory)
-        self._assert_trace_structure_from_writer_call_args(["workflow", "workflow", "llm"])
-
-        LLMObs.disable()
-
-    @run_in_subprocess(env_overrides=bedrock_env_config)
-    def test_llmobs_with_llm_model_bedrock_disabled(self):
-        from langchain.chains import ConversationChain
-        from langchain.memory import ConversationBufferMemory
-        from langchain_community.llms import Bedrock
-
-        from ddtrace import patch
-        from ddtrace.llmobs import LLMObs
-
-        patch(langchain=True)
-        LLMObs.enable(ml_app="<ml-app-name>", integrations_enabled=False, agentless_enabled=True)
-        self._call_bedrock_llm(Bedrock, ConversationChain, ConversationBufferMemory)
-        self._assert_trace_structure_from_writer_call_args(["workflow", "llm"])
-
-        LLMObs.disable()
-
-    @run_in_subprocess(env_overrides=openai_env_config)
-    def test_llmobs_langchain_with_openai_enabled(self):
-        from langchain_openai import OpenAI
-
-        from ddtrace import patch
-        from ddtrace.llmobs import LLMObs
-
-        patch(langchain=True, openai=True)
-        LLMObs.enable(ml_app="<ml-app-name>", integrations_enabled=False, agentless_enabled=True)
-        self._call_openai_llm(OpenAI)
-        self._assert_trace_structure_from_writer_call_args(["workflow", "llm"])
-
-        LLMObs.disable()
-
-    @run_in_subprocess(env_overrides=openai_env_config)
-    def test_llmobs_langchain_with_openai_disabled(self):
-        from langchain_openai import OpenAI
-
-        from ddtrace import patch
-        from ddtrace.llmobs import LLMObs
-
-        patch(langchain=True)
-
-        LLMObs.enable(ml_app="<ml-app-name>", integrations_enabled=False, agentless_enabled=True)
-        self._call_openai_llm(OpenAI)
-        self._assert_trace_structure_from_writer_call_args(["llm"])
-
-        LLMObs.disable()
