+++ tests/contrib/langchain/test_langchain_patch.py
@@ -76,6 +76,9 @@ class TestLangchainPatch(PatchTestCase.Base):
             self.assert_not_wrapped(langchain_openai.OpenAIEmbeddings.embed_documents)
             self.assert_not_wrapped(langchain_pinecone.PineconeVectorStore.similarity_search)
         else:
+            from langchain import embeddings  # noqa: F401
+            from langchain import vectorstores  # noqa: F401
+
             gated_langchain = langchain
             self.assert_not_wrapped(langchain.llms.base.BaseLLM.generate)
             self.assert_not_wrapped(langchain.llms.base.BaseLLM.agenerate)
