+++ tests/snapshots/tests.contrib.langchain.test_langchain.test_openai_sequential_chain_with_multiple_llm_sync.json
@@ -10,13 +10,13 @@
     "error": 0,
     "meta": {
       "_dd.p.dm": "-0",
-      "_dd.p.tid": "6615aa1a00000000",
-      "langchain.request.inputs.input_text": "\\n            I have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\n           ...",
+      "_dd.p.tid": "66566a3300000000",
+      "langchain.request.inputs.input_text": "\\nI have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\nbodies. Does it now fol...",
       "langchain.request.type": "chain",
       "langchain.response.outputs.final_output": "\\nI've come to the conclusion it's true,\\nThere's nothing in the world like me and you.\\nDoes this mean I don't exist? No!\\nFor ...",
-      "langchain.response.outputs.input_text": "\\n            I have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\n           ...",
+      "langchain.response.outputs.input_text": "\\nI have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\nbodies. Does it now fol...",
       "language": "python",
-      "runtime-id": "5bdc2648fed64ec2a49138caab5a0bf0"
+      "runtime-id": "103801afb5f54bf4b14c4af40585cea5"
     },
     "metrics": {
       "_dd.measured": 1,
@@ -27,10 +27,10 @@
       "langchain.tokens.prompt_tokens": 331,
       "langchain.tokens.total_cost": 0.011720000000000001,
       "langchain.tokens.total_tokens": 586,
-      "process_id": 45546
+      "process_id": 69928
     },
-    "duration": 6313000,
-    "start": 1712695834803226000
+    "duration": 8782000,
+    "start": 1716939315772875000
   },
      {
        "name": "langchain.request",
@@ -42,7 +42,7 @@
        "type": "llm",
        "error": 0,
        "meta": {
-         "langchain.request.inputs.input_text": "\\n            I have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\n           ...",
+         "langchain.request.inputs.input_text": "\\nI have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\nbodies. Does it now fol...",
          "langchain.request.prompt": "Paraphrase this text:\\n\\n        {input_text}\\n\\n        Paraphrase: ",
          "langchain.request.type": "chain",
          "langchain.response.outputs.paraphrased_output": "\\nI have come to the conclusion that there is nothing in the world such as the sky, the earth, minds, or bodies. Does this mean ..."
@@ -54,8 +54,8 @@
          "langchain.tokens.total_cost": 0.006240000000000001,
          "langchain.tokens.total_tokens": 312
        },
-       "duration": 2738000,
-       "start": 1712695834803367000
+       "duration": 6067000,
+       "start": 1716939315773033000
      },
         {
           "name": "langchain.request",
@@ -77,7 +77,7 @@
             "langchain.request.openai.parameters.request_timeout": "None",
             "langchain.request.openai.parameters.temperature": "0.7",
             "langchain.request.openai.parameters.top_p": "1",
-            "langchain.request.prompts.0": "Paraphrase this text:\\n\\n        \\n            I have convinced myself that there is absolutely nothing in the world, no sky, no...",
+            "langchain.request.prompts.0": "Paraphrase this text:\\n\\n        \\nI have convinced myself that there is absolutely nothing in the world, no sky, no earth, no m...",
             "langchain.request.provider": "openai",
             "langchain.request.type": "llm",
             "langchain.response.completions.0.finish_reason": "stop",
@@ -91,8 +91,8 @@
             "langchain.tokens.total_cost": 0.006240000000000001,
             "langchain.tokens.total_tokens": 312
           },
-          "duration": 2513000,
-          "start": 1712695834803528000
+          "duration": 5836000,
+          "start": 1716939315773210000
         },
      {
        "name": "langchain.request",
@@ -104,7 +104,7 @@
        "type": "llm",
        "error": 0,
        "meta": {
-         "langchain.request.inputs.input_text": "\\n            I have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\n           ...",
+         "langchain.request.inputs.input_text": "\\nI have convinced myself that there is absolutely nothing in the world, no sky, no earth, no minds, no\\nbodies. Does it now fol...",
          "langchain.request.inputs.paraphrased_output": "\\nI have come to the conclusion that there is nothing in the world such as the sky, the earth, minds, or bodies. Does this mean ...",
          "langchain.request.prompt": "Make this text rhyme:\\n\\n        {paraphrased_output}\\n\\n        Rhyme: ",
          "langchain.request.type": "chain",
@@ -117,8 +117,8 @@
          "langchain.tokens.total_cost": 0.0054800000000000005,
          "langchain.tokens.total_tokens": 274
        },
-       "duration": 3305000,
-       "start": 1712695834806198000
+       "duration": 2486000,
+       "start": 1716939315779141000
      },
         {
           "name": "langchain.request",
@@ -154,6 +154,6 @@
             "langchain.tokens.total_cost": 0.0054800000000000005,
             "langchain.tokens.total_tokens": 274
           },
-          "duration": 3083000,
-          "start": 1712695834806377000
+          "duration": 2254000,
+          "start": 1716939315779332000
         }]]
