+++ tests/snapshots/tests.contrib.langchain.test_langchain_community.test_openai_service_name[None-v1].json
@@ -98,7 +98,7 @@
             "http.method": "POST",
             "http.status_code": "200",
             "http.url": "https://api.openai.com/v1/completions",
-            "http.useragent": "OpenAI/Python 1.12.0",
+            "http.useragent": "OpenAI/Python 1.30.3",
             "out.host": "api.openai.com",
             "peer.service": "api.openai.com",
             "span.kind": "client"
